Read UFS function

This function allows you to read UFS assignment data into R.

read_ufs(file, clean_up = TRUE, load_geometry = FALSE,
  selection_mode = "all_links", step = 1, variables = NULL)

Arguments

file
The file to read.

clean_up
when TRUE removes the VDU, KEY and LPX files

load_geometry
it uses satdb to load the geometry and converts the object into a sf

selection_mode
Define how to query the ufs with the options 'all_links'(default), 'simulation_links' 'simulation_turns' or 'centroid_connectors'.

step
how many records to read by iteration, max is 9, use 1 for stability.